pyproject.toml
[project]
name = "arrhythmia-ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"


dependencies = [
    "matplotlib==3.10.8",
    "mlflow==3.10.0",
    "neurokit2==0.2.12",
    "numpy==2.4.2",
    "pandas==2.3.3",
    "scipy==1.17.0",
    "scikit-learn==1.8.0",
    "joblib==1.5.3",
    "wfdb==4.3.1",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.3",
    "aiosignal==1.4.0",
    "alembic==1.18.4",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "asttokens==3.0.1",
    "attrs==25.4.0",
    "blinker==1.9.0",
    "cachetools==7.0.1",
    "certifi==2026.1.4",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "cloudpickle==3.1.2",
    "colorama==0.4.6",
    "comm==0.2.3",
    "contourpy==1.3.3",
    "cryptography==46.0.5",
    "cycler==0.12.1",
    "databricks-sdk==0.94.0",
    "debugpy==1.8.20",
    "decorator==5.2.1",
    "docker==7.1.0",
    "executing==2.2.1",
    "fastapi==0.135.0",
    "flask==3.1.3",
    "flask-cors==6.0.2",
    "fonttools==4.61.1",
    "frozenlist==1.8.0",
    "fsspec==2026.2.0",
    "gitdb==4.0.12",
    "gitpython==3.1.46",
    "google-auth==2.48.0",
    "graphene==3.4.3",
    "graphql-core==3.2.7",
    "graphql-relay==3.2.0",
    "greenlet==3.3.2",
    "h11==0.16.0",
    "huey==2.6.0",
    "idna==3.11",
    "imbalanced-learn==0.14.1",
    "importlib-metadata==8.7.1",
    "ipykernel==7.2.0",
    "ipython==9.10.0",
    "ipython-pygments-lexers==1.1.1",
    "itsdangerous==2.2.0",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jupyter-client==8.8.0",
    "jupyter-core==5.9.1",
    "kiwisolver==1.4.9",
    "mako==1.3.10",
    "markupsafe==3.0.3",
    "matplotlib-inline==0.2.1",
    "mlflow-skinny==3.10.0",
    "mlflow-tracing==3.10.0",
    "multidict==6.7.1",
    "nest-asyncio==1.6.0",
    "opentelemetry-api==1.39.1",
    "opentelemetry-proto==1.39.1",
    "opentelemetry-sdk==1.39.1",
    "opentelemetry-semantic-conventions==0.60b1",
    "packaging==26.0",
    "parso==0.8.5",
    "pillow==12.1.0",
    "platformdirs==4.5.1",
    "prettytable==3.17.0",
    "prompt-toolkit==3.0.52",
    "propcache==0.4.1",
    "protobuf==6.33.5",
    "psutil==7.2.2",
    "pure-eval==0.2.3",
    "pyarrow==23.0.1",
    "pyasn1==0.6.2",
    "pyasn1-modules==0.4.2",
    "pycparser==3.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pygments==2.19.2",
    "pyparsing==3.3.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "pytz==2025.2",
    "pywavelets==1.9.0",
    "pyyaml==6.0.3",
    "pyzmq==27.1.0",
    "rsa==4.9.1",
    "seaborn==0.13.2",
    "six==1.17.0",
    "sklearn-compat==0.1.5",
    "skops==0.13.0",
    "smmap==5.0.2",
    "soundfile==0.13.1",
    "sqlalchemy==2.0.47",
    "sqlparse==0.5.5",
    "stack-data==0.6.3",
    "starlette==0.52.1",
    "threadpoolctl==3.6.0",
    "tornado==6.5.4",
    "traitlets==5.14.3",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.3",
    "urllib3==2.6.3",
    "uvicorn==0.41.0",
    "waitress==3.0.2",
    "wcwidth==0.6.0",
    "werkzeug==3.1.6",
    "xgboost==3.2.0",
    "yarl==1.22.0",
    "zipp==3.23.0",
    "pdoc>=16.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "hatchling>=1.29.0",
]
